centralize module typing overrides (#1786)
* optimize typings with declaration file
