+++ src/main.ts
@@ -73,74 +73,12 @@ const config: Phaser.Types.Core.GameConfig = {
 
 /**
  * Sets this object's position relative to another object with a given offset
- * @param guideObject {@linkcode Phaser.GameObjects.GameObject} to base the position off of
- * @param x The relative x position
- * @param y The relative y position
  */
-const setPositionRelative = function (guideObject: any, x: number, y: number) {
+const setPositionRelative = function (guideObject: Phaser.GameObjects.GameObject, x: number, y: number) {
   const offsetX = guideObject.width * (-0.5 + (0.5 - guideObject.originX));
   const offsetY = guideObject.height * (-0.5 + (0.5 - guideObject.originY));
   this.setPosition(guideObject.x + offsetX + x, guideObject.y + offsetY + y);
 };
-
-declare module "phaser" {
-	namespace GameObjects {
-		interface Container {
-      /**
-       * Sets this object's position relative to another object with a given offset
-       * @param guideObject {@linkcode Phaser.GameObjects.GameObject} to base the position off of
-       * @param x The relative x position
-       * @param y The relative y position
-       */
-			setPositionRelative(guideObject: any, x: number, y: number): void;
-		}
-		interface Sprite {
-      /**
-       * Sets this object's position relative to another object with a given offset
-       * @param guideObject {@linkcode Phaser.GameObjects.GameObject} to base the position off of
-       * @param x The relative x position
-       * @param y The relative y position
-       */
-			setPositionRelative(guideObject: any, x: number, y: number): void;
-		}
-		interface Image {
-      /**
-       * Sets this object's position relative to another object with a given offset
-       * @param guideObject {@linkcode Phaser.GameObjects.GameObject} to base the position off of
-       * @param x The relative x position
-       * @param y The relative y position
-       */
-			setPositionRelative(guideObject: any, x: number, y: number): void;
-		}
-		interface NineSlice {
-      /**
-       * Sets this object's position relative to another object with a given offset
-       * @param guideObject {@linkcode Phaser.GameObjects.GameObject} to base the position off of
-       * @param x The relative x position
-       * @param y The relative y position
-       */
-			setPositionRelative(guideObject: any, x: number, y: number): void;
-		}
-		interface Text {
-      /**
-       * Sets this object's position relative to another object with a given offset
-       * @param guideObject {@linkcode Phaser.GameObjects.GameObject} to base the position off of
-       * @param x The relative x position
-       * @param y The relative y position
-       */
-			setPositionRelative(guideObject: any, x: number, y: number): void;
-		}
-		interface Rectangle {
-      /**
-       * Sets this object's position relative to another object with a given offset
-       * @param guideObject {@linkcode Phaser.GameObjects.GameObject} to base the position off of
-       * @param x The relative x position
-       * @param y The relative y position
-       */
-			setPositionRelative(guideObject: any, x: number, y: number): void;
-		}
-	}
-}
 
 Phaser.GameObjects.Container.prototype.setPositionRelative = setPositionRelative;
 Phaser.GameObjects.Sprite.prototype.setPositionRelative = setPositionRelative;
